requirements.txt
Flask==3.0.0
Pillow>=10.0.0
img2pdf>=0.5.0
harfbuzz>=4.4.0
freetype-py>=2.3.0

